Rename all route files into lowercase
