+++ src/routes/settings-audio.tsx
@@ -5,7 +5,7 @@ import * as Setting from '../components/Setting';
 import { usePlayerAPI } from '../stores/usePlayerStore';
 
 import useInvalidate, { useInvalidateCallback } from '../hooks/useInvalidate';
-import type { SettingsLoaderData } from './ViewSettings';
+import type { SettingsLoaderData } from './settings';
 
 export default function ViewSettingsAudio() {
   const { config } = useLoaderData() as SettingsLoaderData;
