+++ src/routes/settings.tsx
@@ -7,7 +7,7 @@ import View from '../elements/View';
 import config from '../lib/config';
 import type { LoaderData } from '../types/museeks';
 
-import styles from './ViewSettings.module.css';
+import styles from './settings.module.css';
 
 export default function ViewSettingsView() {
   const match = useMatch('/settings');
